% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitflyer-package.R
\docType{package}
\name{bitflyer}
\alias{bitflyer}
\alias{bitflyer-package}
\title{R Wrapper for bitFlyer's REST API}
\description{
Make it easier to use bitFlyer's REST API (\url{https://lightning.bitflyer.com/docs?lang=en})
}
\author{
\strong{Maintainer}: Nagi Teramo \email{teramonagi@gmail.com}

Authors:
\itemize{
  \item Shinichi Takayanagi \email{shinichi.takayanagi@gmail.com}
}

}
\keyword{internal}
