% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{board}
\alias{board}
\title{Order Book}
\usage{
board(product_code = "BTC_JPY")
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}
}
\description{
Get order book
}
\examples{
\dontrun{
  board()
  board(product_code = "BTC_USD")
}
}
