% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{cancel_child_order}
\alias{cancel_child_order}
\title{Cancel Order}
\usage{
cancel_child_order(
  product_code,
  child_order_id = NULL,
  child_order_acceptance_id = NULL
)
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}

\item{child_order_id}{ID for the child order.}

\item{child_order_acceptance_id}{Expects an ID from \code{\link{send_child_order}}}
}
\description{
Cancel Order
}
