% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{cancel_parent_order}
\alias{cancel_parent_order}
\title{Cancel parent order}
\usage{
cancel_parent_order()
}
\arguments{
\item{parent_order_id}{ID for the canceling order.}

\item{parent_order_acceptance_id}{Expects an ID from Submit New Parent Order.
When specified, the corresponding order will be cancelled.}
}
\description{
Parent orders can be canceled in the same manner as regular orders.
If a parent order is canceled, the placed orders associated with that order will all be canceled.
}
