% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{get_board_state}
\alias{get_board_state}
\title{Orderbook status}
\usage{
get_board_state(product_code = "BTC_JPY")
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}
}
\value{
status: one of the following levels will be displayed
  \itemize{
    \item NORMAL: The exchange is operating.
    \item BUSY: The exchange is experiencing high traffic.
    \item VERY BUSY: The exchange is experiencing very heavy traffic.
    \item SUPER BUSY: The exchange is experiencing extremely heavy traffic. There is a possibility that orders will fail or be processed after a delay.
    \item STOP: The exchange has been stopped. Orders will not be accepted.
 }
}
\description{
Get orderbook status
}
\examples{
\dontrun{
  get_board_state(product_code = "BTC_JPY")
}

}
