% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_child_orders}
\alias{get_child_orders}
\title{List Orders}
\usage{
get_child_orders()
}
\arguments{
\item{product_code}{Please specify a product_code or alias, as obtained from the Market List. Only "BTC_USD" is available for U.S. accounts, and only "BTC_EUR" is available for European accounts.}

\item{count, }{before, after: See Pagination.}

\item{child_order_state}{When specified, return only orders that match the specified value.
You must specify one of the following:
ACTIVE: Return open order
COMPLETED: Return fully completed orders
CANCELED: Return orders that have been cancelled by the customer
EXPIRED: Return order that have been cancelled due to expiry
REJECTED: Return failed orders}

\item{child_order_acceptance_id}{ID for the child order.}

\item{parent_order_id}{If specified, a list of all orders associated with the parent order is obtained.}
}
\description{
List Orders
}
