% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_collateral_accounts}
\alias{get_collateral_accounts}
\title{Get Margin Status}
\usage{
get_collateral_accounts()
}
\description{
Get details of your collateral balances (multiple currencies supported).
}
