% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_executions}
\alias{get_executions}
\title{List Executions}
\usage{
get_executions(
  product_code,
  count = 100,
  before = 0,
  after = 0,
  child_order_id = NULL,
  child_order_acceptance_id = NULL
)
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}

\item{count}{Specifies the number of results. If this is omitted, the value will be 100.}

\item{before}{Obtains data having an id lower than the value specified for this parameter.}

\item{after}{Obtains data having an id higher than the value specified for this parameter.}

\item{child_order_id}{ID for the child order.}

\item{child_order_acceptance_id}{Expects an ID from \code{\link{send_child_order}}}
}
\description{
List Executions
}
\details{
When \code{child_order_id} is specified, a list of stipulations related to the order will be displayed.
When \code{child_order_acceptance_id} is specified, a list of stipulations related to the corresponding order will be displayed.
}
