% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{get_health}
\alias{get_health}
\title{Exchange status}
\usage{
get_health(product_code = "BTC_JPY")
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}
}
\value{
One of the following levels
   NORMAL: The exchange is operating.
   BUSY: The exchange is experiencing high traffic.
   VERY BUSY: The exchange is experiencing heavy traffic.
   SUPER BUSY: The exchange is experiencing extremely heavy traffic.
     There is a possibility that orders will fail or be processed after a delay.
   STOP: The exchange has been stopped. Orders will not be accepted.
}
\description{
Get the current status of the exchange.
}
\examples{
\dontrun{
get_health()
}

}
