% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_positions}
\alias{get_positions}
\title{Get Open Interest Summary}
\usage{
get_positions(..., region = "")
}
\arguments{
\item{product_code}{Currently supports only "FX_BTC_JPY".}
}
\description{
Get Open Interest Summary
}
