% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_trading_commission}
\alias{get_trading_commission}
\title{Get Trading Commission}
\usage{
get_trading_commission(product_code)
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}
}
\description{
Get Trading Commission
}
