% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{markets}
\alias{markets}
\title{Market List}
\usage{
markets(region = "")
}
\arguments{
\item{region}{Default value ("") is Japan.
\itemize{
  \item Japan: ""(Default), "ja", "jpn
  \item Usa: "us", "usa"
  \item Euro: "eu"
}}
}
\description{
Get the list of market
}
\examples{
\dontrun{
  markets()
  markets(region = "us")
}
}
