% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual-template.R
\name{parent_order_acceptance_id}
\alias{parent_order_acceptance_id}
\title{Params parent_order_acceptance_id}
\arguments{
\item{parent_order_acceptance_id}{Expects an ID from \code{\link{send_parent_order}}}
}
\keyword{internal}
