% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual-template.R
\name{parent_order_state}
\alias{parent_order_state}
\title{parent_order_state}
\arguments{
\item{parent_order_state}{When specified, return only orders that match the specified value.
You must specify one of the following:
\itemize{
  \item "ACTIVE": Return open orders
  \item "COMPLETED": Return fully completed orders
  \item "CANCELED": Return orders that have been cancelled by the customer
  \item "EXPIRED": Return order that have been cancelled due to expiry
  \item "REJECTED": Return failed orders
}}
}
\description{
parent_order_state
}
\keyword{internal}
