% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{product_code}
\alias{product_code}
\alias{func_product_code}
\title{Params product_code}
\usage{
func_product_code(product_code = "BTC_JPY")
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}
}
\keyword{internal}
