% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual-template.R
\name{region}
\alias{region}
\title{region}
\arguments{
\item{region}{Default value ("") is Japan.
\itemize{
  \item Japan: ""(Default), "ja", "jpn
  \item Usa: "us", "usa"
  \item Euro: "eu"
}}
}
\description{
region
}
\keyword{internal}
