% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{send_parent_order}
\alias{send_parent_order}
\title{Submit New Parent Order (Special order)}
\usage{
send_parent_order(
  order_method,
  minute_to_expire,
  time_in_force,
  parameters,
  product_code,
  condition_type,
  side,
  size,
  price,
  trigger_price,
  offset
)
}
\arguments{
\item{order_method}{The order method. Please set it to one of the following values. If omitted, the value defaults to "SIMPLE".
\itemize{
  \item "SIMPLE": A special order whereby one order is placed.
  \item "IFD": Conducts an IFD order. In this method, you place two orders at once, and when the first order is completed, the second order is automatically placed.
  \item "OCO": Conducts an OCO order. In this method, you place two orders at one, and when one of the orders is completed, the other order is automatically canceled.
  \item "IFDOCO": Conducts an IFD-OCO order. In this method, once the first order is completed, an OCO order is automatically placed.
}}

\item{minute_to_expire}{Specifies the time until the order expires in minutes. If omitted, the value defaults to 43200 (30 days).}

\item{time_in_force}{Specify any of the following execution conditions - "GTC", "IOC", or "FOK". If omitted, the value defaults to "GTC".}

\item{parameters}{This is an array that specifies the parameters of the order to be placed.
The required length of the array varies depending upon the specified order_method.
If "SIMPLE" has been specified, specify one parameter.
If "IFD" has been specified, specify two parameters. The first parameter is the parameter for the first order placed. The second parameter is the parameter for the order to be placed after the first order is completed.
If "OCO" has been specified, specify two parameters. Two orders are placed simultaneously based on these parameters.
If "IFDOCO" has been specified, specify three parameters. The first parameter is the parameter for the first order placed. After the order is complete, an OCO order is placed with the second and third parameters.
In the parameters, specify an array of objects with the following keys and values.}

\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}

\item{condition_type}{This is the execution condition for the order. Please set it to one of the following values.
 \itemize{
   \item "LIMIT": Limit order.
   \item "MARKET": Market order.
   \item "STOP": Stop order.
   \item "STOP_LIMIT": Stop-limit order.
   \item "TRAIL": Trailing stop order.
}}

\item{side}{For buying orders, specify "BUY", for selling orders, specify "SELL".}

\item{size}{Specify the order quantity.}

\item{price}{Specify the price. This is a required value if condition_type has been set to "LIMIT" or "STOP_LIMIT".}

\item{trigger_price}{Specify the trigger price for a stop order. This is a required value if condition_type has been set to "STOP" or "STOP_LIMIT".}

\item{offset}{Specify the trail width of a trailing stop order as a positive integer. This is a required value if condition_type has been set to "TRAIL".}
}
\description{
Submit New Parent Order (Special order)
}
\details{
It is possible to place orders including logic other than simple limit orders (LIMIT)
and market orders (MARKET). Such orders are handled as parent orders.
By using a special order, it is possible to place orders in response to market conditions or place multiple associated orders.

Please read about the types of special orders and their methods in the bitFlyer Lightning documentation on special orders:
\url{https://lightning.bitflyer.com/docs/specialorder}.
}
