% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{cancel_child_order}
\alias{cancel_child_order}
\title{Cancel Order}
\usage{
cancel_child_order(product_code, child_order_type, side, price, size,
  minute_to_expire = 43200, time_in_force = "GTC")
}
\arguments{
\item{product_code}{Required. The product for the corresponding order. Please specify a product_code or alias, as obtained from the Market List. Only "BTC_USD" is available for U.S. accounts, and only "BTC_EUR" is available for European accounts.}

\item{child_order_id}{ID for the canceling order.}

\item{child_order_acceptance_id}{Expects an ID from Send a New Order. When specified, the corresponding order will be cancelled.}
}
\description{
Cancel Order
}
