% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{cancel_parent_order}
\alias{cancel_parent_order}
\title{Cancel parent order}
\usage{
cancel_parent_order(..., region = "")
}
\description{
Parent orders can be canceled in the same manner as regular orders.
If a parent order is canceled, the placed orders associated with that order will all be canceled.
}
\details{
When \code{parent_order_acceptance_id} is specified, the corresponding order will be cancelled.
}
