% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual-template.R
\name{child_order_state}
\alias{child_order_state}
\title{Params cild_order_state}
\arguments{
\item{child_order_state}{When specified, return only orders that match the specified value.
You must specify one of the following:
ACTIVE: Return open orders
COMPLETED: Return fully completed orders
CANCELED: Return orders that have been cancelled by the customer
EXPIRED: Return order that have been cancelled due to expiry
REJECTED: Return failed orders}
}
\keyword{internal}
