% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{executions}
\alias{executions}
\title{Execution History}
\usage{
executions(product_code = "BTC_JPY", count = 100, before = NA,
  after = NA)
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}

\item{count}{Specifies the number of results. If this is omitted, the value will be 100.}

\item{before}{Obtains data having an id lower than the value specified for this parameter.}

\item{after}{Obtains data having an id higher than the value specified for this parameter.}
}
\description{
Get Execution History
}
\note{
API endpoint \code{/v1/getexecutions} is not implemented because there is the same name endpoint in private API
}
