% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{get_chats}
\alias{get_chats}
\title{Chat}
\usage{
get_chats(...)
}
\arguments{
\item{from_date}{string (e.g. 2018-06-30). This function wil return any new messages after this date.}
}
\description{
Get the history of chats
}
