% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_child_orders}
\alias{get_child_orders}
\title{List Orders}
\usage{
get_child_orders(
  product_code,
  count = 100,
  before = 0,
  after = 0,
  child_order_state = NULL,
  child_order_id = NULL,
  child_order_acceptance_id = NULL,
  parent_order_id = NULL
)
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}

\item{count}{Specifies the number of results. If this is omitted, the value will be 100.}

\item{before}{Obtains data having an id lower than the value specified for this parameter.}

\item{after}{Obtains data having an id higher than the value specified for this parameter.}

\item{child_order_state}{When specified, return only orders that match the specified value.
You must specify one of the following:
\itemize{
  \item "ACTIVE": Return open orders
  \item "COMPLETED": Return fully completed orders
  \item "CANCELED": Return orders that have been cancelled by the customer
  \item "EXPIRED": Return order that have been cancelled due to expiry
  \item "REJECTED": Return failed orders
}}

\item{child_order_id}{ID for the child order.}

\item{child_order_acceptance_id}{Expects an ID from \code{\link{send_child_order}}}

\item{parent_order_id}{The ID of the parent order in question.}
}
\description{
List Orders
}
\details{
If \code{parent_order_id} is specified, a list of all orders associated with the parent order is obtained.
}
