% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_executions}
\alias{get_executions}
\title{List Executions}
\usage{
get_executions()
}
\description{
List Executions
}
\details{
When \code{child_order_id} is specified, a list of stipulations related to the order will be displayed.
When \code{child_order_acceptance_id} is specified, a list of stipulations related to the corresponding order will be displayed.
}
