% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_parent_order}
\alias{get_parent_order}
\title{Get Parent Order Details}
\usage{
get_parent_order(parent_order_id, parent_order_acceptance_id)
}
\arguments{
\item{parent_order_id}{The ID of the parent order in question.}

\item{parent_order_acceptance_id}{The acceptance ID for the API to place a new parent order.
If specified, it returns the details of the parent order in question.}
}
\description{
Get Parent Order Details
}
