% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_parent_order}
\alias{get_parent_order}
\title{Get Parent Order Details}
\usage{
get_parent_order(parent_order_id = NULL, parent_order_acceptance_id = NULL)
}
\arguments{
\item{parent_order_id}{The ID of the parent order in question.}

\item{parent_order_acceptance_id}{Expects an ID from \code{\link{send_parent_order}}}
}
\description{
Get Parent Order Details
Please specify either parent_order_id or parent_order_acceptance_id.
}
