% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_parent_orders}
\alias{get_parent_orders}
\title{List Parent Orders}
\usage{
get_parent_orders(
  product_code,
  count = 100,
  before = 0,
  after = 0,
  parent_order_state = NULL
)
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}

\item{count}{Specifies the number of results. If this is omitted, the value will be 100.}

\item{before}{Obtains data having an id lower than the value specified for this parameter.}

\item{after}{Obtains data having an id higher than the value specified for this parameter.}

\item{parent_order_state}{When specified, return only orders that match the specified value.
You must specify one of the following:
\itemize{
  \item "ACTIVE": Return open orders
  \item "COMPLETED": Return fully completed orders
  \item "CANCELED": Return orders that have been cancelled by the customer
  \item "EXPIRED": Return order that have been cancelled due to expiry
  \item "REJECTED": Return failed orders
}}
}
\description{
List Parent Orders
}
