% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{get_parent_orders}
\alias{get_parent_orders}
\title{List Parent Orders}
\usage{
get_parent_orders()
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}

\item{child_order_state}{When specified, return only orders that match the specified value.
You must specify one of the following:
ACTIVE: Return open orders
COMPLETED: Return fully completed orders
CANCELED: Return orders that have been cancelled by the customer
EXPIRED: Return order that have been cancelled due to expiry
REJECTED: Return failed orders}
}
\description{
List Parent Orders
}
