% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{markets}
\alias{markets}
\alias{get_markets}
\title{Market List}
\usage{
markets(region = "")

get_markets(region = "")
}
\arguments{
\item{region}{empty string ("") or "usa" or "eu"}
}
\description{
Get the list of market
}
