% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{send_child_order}
\alias{send_child_order}
\title{Send a New Order}
\usage{
send_child_order(
  product_code,
  child_order_type,
  side,
  price,
  size,
  minute_to_expire = 43200,
  time_in_force = "GTC"
)
}
\arguments{
\item{product_code}{A product_code or alias, as obtained from the Market List.
If omitted, the value is set to "BTC_JPY".
Only "BTC_USD" is available for U.S. accounts,
and only "BTC_EUR" is available for European accounts.}

\item{child_order_type}{For limit orders, it will be "LIMIT". For market orders, "MARKET".}

\item{side}{For buy orders, "BUY". For sell orders, "SELL".}

\item{price}{Specify the price. This is a required value if child_order_type has been set to "LIMIT".}

\item{size}{Specify the order quantity.}

\item{minute_to_expire}{Specify the time in minutes until the expiration time.
If omitted, the value will be 43200 (30 days).}

\item{time_in_force}{Specify any of the following execution conditions - "GTC", "IOC", or "FOK".
If omitted, the value defaults to "GTC".}
}
\value{
child_order_acceptance_id.
  To specify the order to return, please use this instead of child_order_id.
  Please confirm the item is either Cancel Order or Obtain Execution List.
}
\description{
Send a New Order
}
