% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{send_parent_order}
\alias{send_parent_order}
\title{Submit New Parent Order (Special order)}
\usage{
send_parent_order()
}
\description{
It is possible to place orders including logic other than simple limit orders (LIMIT)
and market orders (MARKET). Such orders are handled as parent orders.
By using a special order, it is possible to place orders in response to market conditions or place multiple associated orders.
}
\details{
Please read about the types of special orders and their methods in the bitFlyer Lightning documentation on special orders:
\url{https://lightning.bitflyer.com/docs/specialorder}.
}
