% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{withdraw}
\alias{withdraw}
\title{Withdrawing Funds}
\usage{
withdraw(..., region = "")
}
\arguments{
\item{currency_code}{Required. Currently only compatible with "JPY" for Japanese accounts,
"USD" for U.S. accounts, and "EUR" for European accounts.}

\item{bank_account_id}{Required. Specify id of the bank account.}

\item{amount}{Required. This is the amount that you are canceling.}
}
\description{
Withdrawing Funds
}
